+++ src/Services/Admin/Contacts/ContactsFetcher.php
@@ -2,51 +2,54 @@
 
 declare(strict_types=1);
 
-namespace B24io\Loyalty\SDK\Services\Admin\Cards;
+namespace B24io\Loyalty\SDK\Services\Admin\Contacts;
 
 use B24io\Loyalty\SDK\Common\Requests\ItemsOrder;
-use B24io\Loyalty\SDK\Common\Result\Cards\CardItemResult;
+use B24io\Loyalty\SDK\Common\Result\Contacts\ContactItemResult;
 use B24io\Loyalty\SDK\Core\Exceptions\BaseException;
 use Generator;
 use Psr\Log\LoggerInterface;
 
-class CardsFetcher
+readonly class ContactsFetcher
 {
     public function __construct(
-        private Cards           $cards,
+        private Contacts        $contacts,
         private LoggerInterface $logger
     )
     {
     }
 
     /**
-     * @return Generator<CardItemResult>
+     * @return Generator<ContactItemResult>
      * @throws BaseException
      */
-    public function list(?ItemsOrder $order = null): Generator
+    public function list(
+        ?ContactsFilter $filter = null,
+        ?ItemsOrder     $order = null,
+    ): Generator
     {
-        $res = $this->cards->list();
+        $res = $this->contacts->list($filter);
 
         $pages = $res->getCoreResponse()->getResponseData()->pagination->pages;
-        $this->logger->debug('CardsFetcher.list.start', [
+        $this->logger->debug('ContactsFetcher.list.start', [
             'total' => $res->getCoreResponse()->getResponseData()->pagination->total,
             'pages' => $pages
         ]);
 
-        $cardCnt = 0;
+        $contactCnt = 0;
         for ($i = 1; $i <= $pages; $i++) {
-            $res = $this->cards->list($order, $i);
+            $res = $this->contacts->list($filter, $order, $i);
             $this->logger->info('CardsFetcher.list.pageItem', [
                 'page' => $res->getCoreResponse()->getResponseData()->pagination->page,
-                'cardCnt' => $cardCnt,
+                'contactCnt' => $contactCnt,
                 'queryDuration' => $res->getCoreResponse()->getResponseData()->metadata->duration
             ]);
 
-            foreach ($res->getCards() as $card) {
-                $cardCnt++;
-                yield $cardCnt => $card;
+            foreach ($res->getContacts() as $contact) {
+                $contactCnt++;
+                yield $contactCnt => $contact;
             }
         }
-        $this->logger->debug('CardsFetcher.list.finish');
+        $this->logger->debug('ContactsFetcher.list.finish');
     }
 }